split up `CollectionAddAll` check in preparation for #505
