+++ autograder-core/src/main/java/de/firemage/autograder/core/check/api/SequentialAddAll.java
@@ -5,17 +5,15 @@ import de.firemage.autograder.core.ProblemType;
 import de.firemage.autograder.core.check.ExecutableCheck;
 import de.firemage.autograder.core.check.api.UseEnumValues.CtEnumFieldRead;
 import de.firemage.autograder.core.integrated.IntegratedCheck;
+import de.firemage.autograder.core.integrated.MethodUtil;
 import de.firemage.autograder.core.integrated.StatementUtil;
 import de.firemage.autograder.core.integrated.StaticAnalysis;
-import de.firemage.autograder.core.integrated.MethodUtil;
-import de.firemage.autograder.core.integrated.TypeUtil;
+import spoon.processing.AbstractProcessor;
 import spoon.reflect.code.CtBlock;
 import spoon.reflect.code.CtExpression;
-import spoon.reflect.code.CtForEach;
 import spoon.reflect.code.CtInvocation;
 import spoon.reflect.code.CtStatement;
 import spoon.reflect.code.CtVariableAccess;
-import spoon.reflect.code.CtVariableRead;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.CtEnum;
 import spoon.reflect.declaration.CtEnumValue;
@@ -25,7 +23,6 @@ import spoon.reflect.reference.CtExecutableReference;
 import spoon.reflect.reference.CtTypeParameterReference;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.reflect.reference.CtVariableReference;
-import spoon.reflect.visitor.CtScanner;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,11 +32,10 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @ExecutableCheck(reportedProblems = {
-    ProblemType.COLLECTION_ADD_ALL,
-    ProblemType.COMMON_REIMPLEMENTATION_ADD_ENUM_VALUES,
-    ProblemType.COMMON_REIMPLEMENTATION_ADD_ALL
+    ProblemType.SEQUENTIAL_ADD_ALL,
+    ProblemType.USE_ENUM_VALUES,
 })
-public class CollectionAddAll extends IntegratedCheck {
+public class SequentialAddAll extends IntegratedCheck {
     private static final int MIN_ADD_ALL_SIZE = 4;
 
     private record AddInvocation(
@@ -71,7 +67,7 @@ public class CollectionAddAll extends IntegratedCheck {
     }
 
     private static <T> boolean isOrderedCollection(CtTypeReference<T> ctTypeReference) {
-        return Stream.of(java.util.List.class)
+        return Stream.of(List.class)
             .map(ctClass -> ctTypeReference.getFactory().createCtTypeReference(ctClass))
             .anyMatch(ctTypeReference::isSubtypeOf);
     }
@@ -115,7 +111,7 @@ public class CollectionAddAll extends IntegratedCheck {
                         )
                     )
                 ),
-                ProblemType.COMMON_REIMPLEMENTATION_ADD_ENUM_VALUES
+                ProblemType.USE_ENUM_VALUES
             );
             return;
         }
@@ -133,7 +129,7 @@ public class CollectionAddAll extends IntegratedCheck {
                     "suggestion", "%s.addAll(%s)".formatted(ctVariable.getSimpleName(), values)
                 )
             ),
-            ProblemType.COLLECTION_ADD_ALL
+            ProblemType.SEQUENTIAL_ADD_ALL
         );
     }
 
@@ -178,79 +174,16 @@ public class CollectionAddAll extends IntegratedCheck {
         }
     }
 
-    private void checkAddAll(CtForEach ctFor) {
-        List<CtStatement> statements = StatementUtil.getEffectiveStatements(ctFor.getBody());
-        if (statements.size() != 1) {
-            return;
-        }
-
-        if (statements.get(0) instanceof CtInvocation<?> ctInvocation
-            && TypeUtil.isSubtypeOf(ctInvocation.getTarget().getType(), java.util.Collection.class)
-            && MethodUtil.isSignatureEqualTo(ctInvocation.getExecutable(), boolean.class, "add", Object.class)
-            && ctInvocation.getExecutable().getParameters().size() == 1
-            // ensure that the add argument simply accesses the for variable:
-            // for (T t : array) {
-            //     collection.add(t);
-            // }
-            && ctInvocation.getArguments().get(0) instanceof CtVariableRead<?> ctVariableRead
-            && ctVariableRead.getVariable().equals(ctFor.getVariable().getReference())) {
-
-            // allow explicit casting
-            if (!ctInvocation.getArguments().get(0).getTypeCasts().isEmpty()) {
-                return;
-            }
-
-            // handle edge case where the variable is implicitly cast in the invocation (Character in List, but char in iterable)
-            List<CtTypeReference<?>> actualTypeArguments = ctInvocation.getTarget().getType().getActualTypeArguments();
-            if (!actualTypeArguments.isEmpty() && !ctFor.getVariable().getType().equals(actualTypeArguments.get(0))) {
-                return;
-            }
-
-            String addAllArg = ctFor.getExpression().toString();
-            if (ctFor.getExpression().getType().isArray()) {
-                addAllArg = "Arrays.asList(%s)".formatted(addAllArg);
-            }
-
-
-            this.addLocalProblem(
-                ctFor,
-                new LocalizedMessage(
-                    "common-reimplementation",
-                    Map.of(
-                        "suggestion", "%s.addAll(%s)".formatted(
-                            ctInvocation.getTarget(),
-                            addAllArg
-                        )
-                    )
-                ),
-                ProblemType.COMMON_REIMPLEMENTATION_ADD_ALL
-            );
-        }
-    }
-
     @Override
     protected void check(StaticAnalysis staticAnalysis) {
-        staticAnalysis.getModel().getRootPackage().accept(new CtScanner() {
+        staticAnalysis.processWith(new AbstractProcessor<CtBlock<?>>() {
             @Override
-            public <T> void visitCtBlock(CtBlock<T> ctBlock) {
+            public void process(CtBlock<?> ctBlock) {
                 if (ctBlock.isImplicit() || !ctBlock.getPosition().isValidPosition()) {
-                    super.visitCtBlock(ctBlock);
                     return;
                 }
 
                 checkAddAll(ctBlock);
-                super.visitCtBlock(ctBlock);
-            }
-
-            @Override
-            public void visitCtForEach(CtForEach ctFor) {
-                if (ctFor.isImplicit() || !ctFor.getPosition().isValidPosition()) {
-                    super.visitCtForEach(ctFor);
-                    return;
-                }
-
-                checkAddAll(ctFor);
-                super.visitCtForEach(ctFor);
             }
         });
     }
