+++ sample_config.yaml
@@ -123,14 +123,14 @@ problemsToReport:
   - CHAR_RANGE
   - UNUSED_CODE_ELEMENT_PRIVATE
   - COMMON_REIMPLEMENTATION_MAX_MIN
-  - COMMON_REIMPLEMENTATION_ADD_ALL
+  - FOR_LOOP_CAN_BE_INVOCATION
   - COMMON_REIMPLEMENTATION_ARRAYS_FILL
   - EXCEPTION_SHOULD_NEVER_BE_CAUGHT
   - USE_DIFFERENT_VISIBILITY_PUBLIC_FIELD
   - MULTI_THREADING
   - BINARY_OPERATOR_ON_BOOLEAN
   - TODO_COMMENT
-  - COMMON_REIMPLEMENTATION_ADD_ENUM_VALUES
+  - USE_ENUM_VALUES
   - COMMON_REIMPLEMENTATION_SUBLIST
   - REDUNDANT_MODIFIER_VISIBILITY_ENUM_CONSTRUCTOR
   - TOO_MANY_EXCEPTIONS
@@ -149,7 +149,7 @@ problemsToReport:
   - COMMON_REIMPLEMENTATION_ITERABLE_DUPLICATES
   - COLLECTIONS_N_COPIES
   - REDUNDANT_ELSE
-  - COLLECTION_ADD_ALL
+  - SEQUENTIAL_ADD_ALL
   - USE_MODULO_OPERATOR
   - NUMBER_FORMAT_EXCEPTION_IGNORED
   - REDUNDANT_ASSIGNMENT
